% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_required_credits}
\alias{calculate_required_credits}
\title{Calculate Required API Credits}
\usage{
calculate_required_credits(n_urls, report_type = 1)
}
\arguments{
\item{n_urls}{Integer. Number of URLs to analyze}

\item{report_type}{Integer. WAVE report type (1-4)}
}
\value{
Integer value of required credits
}
\description{
Calculates total API credits needed for analysis
}
\keyword{internal}
