% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_generation.R
\name{create_html_table}
\alias{create_html_table}
\title{Create HTML Table}
\usage{
create_html_table(df)
}
\arguments{
\item{df}{Data frame to convert}
}
\value{
Character string containing HTML table
}
\description{
Internal function to convert data frame to HTML table
}
\keyword{internal}
