% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_report_filename}
\alias{generate_report_filename}
\title{Generate WAVE Report Filename}
\usage{
generate_report_filename(base_name, type, ext = "png")
}
\arguments{
\item{base_name}{Character string. Base name for the file}

\item{type}{Character string. Type of report}

\item{ext}{Character string. File extension}
}
\value{
Character string containing formatted filename
}
\description{
Creates a standardized filename for WAVE reports
}
\keyword{internal}
