% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_wave_params}
\alias{validate_wave_params}
\title{Validate WAVE Parameters}
\usage{
validate_wave_params(input, api_key = NULL, report_type = 1)
}
\arguments{
\item{input}{Character vector. URLs or file paths}

\item{api_key}{Character string. WAVE API key}

\item{report_type}{Integer. WAVE report type}
}
\value{
Logical TRUE if valid, stops with error if invalid
}
\description{
Validates input parameters for WAVE API calls
}
\keyword{internal}
